User Story 22. All Unique Ids.
wrote a user story to check whether all the ids of Individuals and families have unique ids. Since database is used to store data, id column is declared as primary key. Hence by default no duplicate ids will be inserted. However, for the purpose of checking, wrote the respective test cases for the same which validated the ids of Individuals and Families. The test case file is named as TESTUS22.txt